[EQLMatch] test match.
