+++ test/test_eql/test_core/test_queries.py
@@ -16,9 +16,7 @@ from krrood.entity_query_language.entity import (
     or_,
     a,
     exists,
-    flatten,
-    match,
-    entity_matching,
+    flatten
 )
 from krrood.entity_query_language.failures import (
     MultipleSolutionFound,
@@ -31,7 +29,6 @@ from krrood.entity_query_language.predicate import (
     symbolic_function,
     Predicate,
 )
-from krrood.entity_query_language.symbol_graph import SymbolGraph
 from krrood.entity_query_language.result_quantification_constraint import (
     ResultQuantificationConstraint,
     Exactly,
@@ -748,34 +745,3 @@ def test_quantified_query(handles_and_containers_world):
         list(get_quantified_query(Exactly(2)).evaluate())
     with pytest.raises(LessThanExpectedNumberOfSolutions):
         list(get_quantified_query(Exactly(4)).evaluate())
-
-
-def test_match(handles_and_containers_world):
-    world = handles_and_containers_world
-
-    fixed_connection_query = the(
-        entity_matching(FixedConnection, world.connections)(
-            parent=match(Container)(name="Container1"),
-            child=match(Handle)(name="Handle1"),
-        )
-    )
-
-    fixed_connection_query_manual = the(
-        entity(
-            fc := let(FixedConnection, domain=None),
-            HasType(fc.parent, Container),
-            HasType(fc.child, Handle),
-            fc.parent.name == "Container1",
-            fc.child.name == "Handle1",
-        )
-    )
-
-    assert fixed_connection_query == fixed_connection_query_manual
-
-    fixed_connection_query.visualize()
-
-    fixed_connection = fixed_connection_query.evaluate()
-    assert isinstance(fixed_connection, FixedConnection)
-    assert fixed_connection.parent.name == "Container1"
-    assert isinstance(fixed_connection.child, Handle)
-    assert fixed_connection.child.name == "Handle1"
